requirements.txt
absl-py==0.15.0
adafruit-ampy==1.1.0
Adafruit-Blinka==8.45.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.71.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.2.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
binho-host-adapter==0.1.6
bitarray==2.9.2
bitstring==4.2.3
black==24.10.0
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docutils==0.21.2
durationpy==0.9
ecdsa==0.19.0
esptool==4.7.0
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.15.4
filetype==1.2.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.5.4
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.7
imageio==2.36.1
imgviz==1.7.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
intelhex==2.3.0
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.3.3
kiwisolver==1.4.5
kubernetes==31.0.0
labelme==5.5.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.147
lazy_loader==0.4
libclang==18.1.1
llama-cloud==0.1.6
llama-index==0.12.6
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.6
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.11
llama-index-multi-modal-llms-openai==0.4.1
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.18
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy==1.10.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.16.0
natsort==8.4.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.58.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.12
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
pipkin==2.1b1
platformdirs==4.2.2
posthog==3.7.4
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pyftdi==0.55.4
Pygments==2.18.0
pylance==0.10.18
pylint==3.2.2
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
pyserial==3.5
PySimpleGUI==5.0.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyusb==1.2.1
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.2.0
QtPy==2.4.2
reedsolo==1.7.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roboflow==1.1.49
rpds-py==0.22.3
rsa==4.9
scikit-image==0.25.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
streamlit-js-eval==0.1.7
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
thonny==4.1.4
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.5
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.35
ultralytics-thop==2.0.12
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.32.1
watchdog==6.0.0
watchfiles==1.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.18.3
zipp==3.21.0
